// minutes/frontend/static_src/src/api/models/TokenSet.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Minutes API V1
 * django-minutes API specifications
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TokenSet
 */
export interface TokenSet {
    /**
     * 
     * @type {string}
     * @memberof TokenSet
     */
    authTokenKey: string;
    /**
     * 
     * @type {Date}
     * @memberof TokenSet
     */
    authTokenExpires: Date;
    /**
     * 
     * @type {string}
     * @memberof TokenSet
     */
    refreshTokenKey: string;
    /**
     * 
     * @type {Date}
     * @memberof TokenSet
     */
    refreshTokenExpires: Date;
}

export function TokenSetFromJSON(json: any): TokenSet {
    return TokenSetFromJSONTyped(json, false);
}

export function TokenSetFromJSONTyped(json: any, ignoreDiscriminator: boolean): TokenSet {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'authTokenKey': json['auth_token_key'],
        'authTokenExpires': (new Date(json['auth_token_expires'])),
        'refreshTokenKey': json['refresh_token_key'],
        'refreshTokenExpires': (new Date(json['refresh_token_expires'])),
    };
}

export function TokenSetToJSON(value?: TokenSet | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'auth_token_key': value.authTokenKey,
        'auth_token_expires': (value.authTokenExpires.toISOString()),
        'refresh_token_key': value.refreshTokenKey,
        'refresh_token_expires': (value.refreshTokenExpires.toISOString()),
    };
}